requirements.txt
pandas==1.5.3
plotly==5.13.1
matplotlib==3.7.1
numpy==1.22.4
streamlit==1.22.0
google-cloud-bigquery==3.11.4
protobuf==3.20.3
db-dtypes